WIP: Initial support for ZFS on FreeBSD
